refactor: http server
